chatbox in lobby and countdown
