+++ frontend/app/game.js
@@ -1,5 +1,6 @@
 
 import { createElement, store } from '../framework/framework.js';
+import { renderChat } from './chat.js';
 
 // WebSocket message sender
 const sendMsg = (msg) => {
@@ -203,66 +204,7 @@ function renderGameGrid(panel, players, powerups) {
     );
 }
 
-// Render the chat area
-function renderChat(messages) {
-    const { playerId } = store.getState();
 
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        const input = e.target.elements.message;
-        if (input.value) {
-            sendMsg({ msgType: 'c', content: input.value });
-            input.value = '';
-        }
-    };
-
-    const renderMessage = (msg) => {
-        const { playerIndex } = store.getState();
-        const isSent = msg.senderIndex === playerIndex;
-        const bubbleClass = isSent ? 'message-bubble sent' : 'message-bubble received';
-        const sender = isSent ? 'You' : msg.player;
-        const timestamp = new Date().toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', hour12: false });
-
-        return createElement('div', { class: 'chat-message' },
-            createElement('div', { class: bubbleClass },
-                createElement('div', { class: 'message-sender', style: `color: ${msg.color}` }, sender),
-                createElement('div', { class: 'message-content' }, msg.message),
-                createElement('div', { class: 'message-timestamp' }, timestamp)
-            )
-        );
-    };
-
-    return createElement('div', { class: 'game-chat' },
-        createElement('div', { class: 'resize-handle', onmousedown: onMouseDown }),
-        createElement('div', { class: 'chat-header' }, 'Game Chat'),
-        createElement('div', { class: 'chat-messages' },
-            ...messages.map(renderMessage)
-        ),
-        createElement('form', { class: 'chat-input-form', onsubmit: handleSubmit },
-            createElement('input', { type: 'text', name: 'message', placeholder: 'Type a message...' }),
-            createElement('button', { type: 'submit' }, '➤')
-        )
-    );
-}
-
-function onMouseDown(e) {
-    e.preventDefault();
-    document.addEventListener('mousemove', onMouseMove);
-    document.addEventListener('mouseup', onMouseUp);
-}
-
-function onMouseMove(e) {
-    const chat = document.querySelector('.game-chat');
-    if (chat) {
-        const newWidth = window.innerWidth - e.clientX;
-        chat.style.width = `${newWidth}px`;
-    }
-}
-
-function onMouseUp() {
-    document.removeEventListener('mousemove', onMouseMove);
-    document.removeEventListener('mouseup', onMouseUp);
-}
 
 // Main Game component
 export default function Game() {
@@ -343,7 +285,8 @@ export default function Game() {
                 createElement('div', { class: 'ice-image ice4-container' },
                     createElement('img', { src: '/public/ice4.png', class: 'ice4-image' }),
                     createElement('div', { class: 'penguin face-r' })
-                )
+                ),
+                renderChat(chatMessages || [])
             );
     }
