+++ frontend/app/modal.css
@@ -19,11 +19,10 @@ body {
 
 .modal {
     position: fixed;
-    z-index: 1000;
+    top: 10vh;
     left: 0;
-    top: 0;
+    z-index: 1000;
     width: 100%;
-    height: 100%;
     display: flex;
     align-items: center;
     justify-content: center;
